fixed comments and changes according to cron.
