+++ routes/redis.js
@@ -11,9 +11,9 @@ router.post('/flush', function (req, res) {
     });
     client.flushdb(function (err, succeeded) {
         if (err) {
-            res.json({err: err, msg: "the redis data is not deleted successfully"});
+            oops(res, err);
         } else {
-            res.json({status: succeeded, msg: "the redis data is deleted successfully"});
+            success(res, 1, succeeded);
         }
     });
 });
@@ -24,13 +24,13 @@ router.post('/set_status', function (req, res) {
     var status = req.body.status;
     dtabase.find({APP_ID: APP_ID}, function (err, result) {
         if (err) {
-            success(res, 0, err);
+            oops(res, err);
         } else if (status == 'enabled' || status == 'disabled') {
             dtabase.update({APP_ID: APP_ID}, {
                 status: status,
             }, function (error) {
                 if (error) {
-                    success(res, 0, error);
+                    oops(res, error);
                 } else {
                     success(res, 1, 'the status is successfully changed to ' + status);
                 }
@@ -46,7 +46,7 @@ router.post('/check_status', function (req, res) {
     var APP_ID = req.headers.app_id;
     dtabase.find({APP_ID: APP_ID}, function (err, result) {
         if (err) {
-            success(res, 0, err);
+            oops(res, err);
         } else {
             success(res, 1, result[0].status);
         }
